create databaseList constants file, to store the column configs constants; refactor everywhere to use them
