+++ redisinsight/ui/src/slices/interfaces/instances.ts
@@ -525,12 +525,3 @@ export enum InstanceType {
   AWSElasticache = 'AWS Elasticache',
   Sentinel = 'Sentinel',
 }
-
-export enum DatabaseListColumn {
-  Name = 'name',
-  Host = 'host',
-  ConnectionType = 'connectionType',
-  Modules = 'modules',
-  LastConnection = 'lastConnection',
-  Controls = 'controls',
-}
